refactor: moved InvalidUpdateError to update command
